requirements.txt
telethon
firebase-admin
python-dotenv